add support for compression/decompression
For now we implement compression at the oci.Image.<method> level, so that
different blobs can be compressed with different algorithms (e.g. perhaps
we don't want to compress metadata, but do want to compress file contents).

Perhaps we should parameterize this better (e.g. "metadata compression
algorithm" and "file compression algorithm"), but for now let's just do
this and see how many times we get it wrong and if/when we want to make
things different.

We implement the "noop" compression algorithm as part of the compression
package, so this should be a noop in functionality. However, we'll soon
have a zstd implementation...

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>