flesh out more OCI support
Previously, we were emitting puzzlefs blobs, but not emitting an OCI image
index with a tag to be able to find the rootfs again. Let's do that.

We also hard code several constants here r.e. puzzlefs, so that puzzlefs
images will not be confused with actual upstream OCI images.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>