reader: change read API to take a buffer as an argument
if the user already has a buffer in hand, they don't want the read() api to
create another buffer only to copy into the buffer they already had. let's
refactor this so that the user allocates the buffer, and the data is copied
in. we also drop some of the casts here.

the motivation for this is an io::Read implementation, which passes a
user-owned buffer in to avoid this copy.

our one user of this API (fuse) can just allocate its buffer as necessary.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>