*: slightly fix error handling
While this does not introduce stack traces to our results (which would be
very handy), it does clean up a lot of terrible casting and wrapper cruft,
duplicate error types, etc.

The way we do this is by just having one WireFormatError, instead of a
bunch of different errors in each package. This means WireFormatError has
to do more, but it greatly simplifies a lot of other code.

We leave stack traces as a future exercise :)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>