builder: add some experiments about FCDC stabilization
Not that we need to be running this code on every commit, but it is useful
to have around and it's not very expensive to run.

At least it is an experimental demonstration of the fact that fastcdc can
stabilize fairly quickly. There's probably some theoretical guarantees one
could derive, but I'm bad at math.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>