builder: introduce simple delta generation
This introduces a very dumb delta generation algorithm, which just re-emits
the entire metadata layer (i.e. it does not omit inodes from a layer if
they are identical to how they were before). However, it does produce
whiteouts if a file is deleted in the new rootfs. For now, we use BSD-style
whiteout inode mode, hopefully there's not a problem with this (perhaps we
should switch to linux kernel style character 0 0 whiteouts for
unprivileged use, but ideally nobody will be extracting themselves and just
use this library).

In order to figure out whether things have been deleted, we change the
"build the FS" loop slightly: we iterate only over directories, and do a
delta detection on a per-directory basis. Then in an inner loop for each
directory, we process its children for file contents, etc. It would be
potentially cleaner to implement some generic sorted-iterator delta thingy
that we could just diff over the existing rootfs and the puzzlefs (which
might also be useful for FS/FS delta in testing), but this works for now.

Obvious next steps are:

1. only emit things in a layer that have changed
2. more testing (for whiteouts, for the above, for a real fs)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>