Fix missing blob error
The error occurs because the code assumes that all the chunks are
generated for all the previous files. This is not the case, since the
chunker splits up only the first part of the buffer into chunks and
there is a leftover part that will be chunked together with the next
files (or if there are no next files, the chunker will be forced to
chunk the leftover).

Since the buffer containing the concatenated files will always be bigger
than the concatenated generated chunks, the new approach is to iterate
through the generated chunks instead of the previous files. Since every
generated chunk will have an associated file or a list of files, the
code will not run into a similar 'missing previous file' error.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>