Log messages to syslog when daemonizing the mount command
The gist of this solution is: create a pipe and redirect the daemon's
stdout and stderr to the writer end of this pipe. The fuse loop is
spawned in a separate thread and any output will be captured by yet
another thread which only reads from the pipe and writes to syslog.

When the filesystem is unmounted, the fuse thread will terminate and it
will notify the main thread of the daemon. This will close the writing
ends of the pipe (stdout and stderr), causing the syslog thread to exit.
Then it will wait for this thread to finish and finally the daemon will
exit.

Use https://github.com/ariel-miculas/daemonize since the daemonize crate
doesn't support redirecting stdout and stderr to pipes. See:
https://github.com/knsd/daemonize/issues/49

Also pin the time crate to version 0.1 because the later versions
deprecate a function which we use.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>