Fix buffer overread in file_read
Ensure we take the offset into account when reading from a file chunk,
otherwise we might end up leaking another file's data, since the last
chunk in the list of chunks can contain data from other files. This
issue was observed while developing the kernel driver. The userspace
implementation might not encounter it due to the way in which fuse
handles the read operations.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>