Enable puzzlefs to be used as a MOUNT_HELPER in the LXC OCI template
The mount-helper is called like this:
mount-helper mount --writable <oci_dir>:<oci_name> <mountpoint>
mount-helper umount <mountpoint>

I've added the support for the `--writable` flag that creates a writable
overlay on top of a PuzzleFS mountpoint (used as a lowerdir). This
requires root privileges.

I've also added the support for the `umount` subcommand which detects
whether the puzzlefs filesystem was mounted as a read-only fuse driver
or whether an overlay was created on top of it, and handles the
unmounting accordingly. Unmounting an overlay requires root privileges.

I've implemented the changes on the LXC side in [1].

[1] https://github.com/lxc/lxc/pull/4483

Signed-off-by: Ariel Miculas-Trif <amiculas@cisco.com>